server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pycryptodome==3.20.0
python-dotenv==1.0.0
werkzeug==2.3.7
